+++ packages/doctor/src/index.ts
@@ -1,3 +1,4 @@
+import * as path from 'path';
 import getFiles from './getFiles';
 import Analyzer from './Analyzer';
 import Scanner from './Scanner';
@@ -27,7 +28,7 @@ class Doctor {
   }
 
   scan(directory: string, options?: IScanOptions): Promise<IScannerReports> {
-    return this.scanner.scan(directory, options);
+    return this.scanner.scan(path.isAbsolute(directory) ? directory : path.join(process.cwd(), directory), options);
   }
 
   analyse(directory: string) {